Add files via upload
Using Long Short-Term Memory (LSTM) to Predict Tesla Stock Price

Long Short-Term Memory (LSTM) is a type of recurrent neural network (RNN) that is well-suited for predicting time series data. It can learn long-term dependencies in the data, which makes it a powerful tool for forecasting stock prices.
In this study, we used LSTM to predict the stock price of Tesla. We used historical stock prices as input to the LSTM model, and the model learned to predict the next day's stock price based on the historical data.
The results of our study showed that LSTM can be used to make accurate predictions of Tesla stock price.
This study suggests that LSTM can be a valuable tool for investors who are looking to predict the future price of Tesla stock. However, it is important to note that no model can perfectly predict the future, and there is always some degree of uncertainty in any stock price prediction.
Bharat Intern #datascienceinternship